+++ country/lib/country.dart
@@ -1,4 +1,3 @@
 library country;
 
 export 'src/country.dart';
-export 'src/country.g.dart';