+++ flutter_country/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.0
+
+* updated country to '1.0.0'
+
 ## 0.0.2
 
 * upgrade country to '0.0.4'